FIXED (NEL-3): major part of compiler warnings regarding non-virtual destructors
git-svn-id: svn://svn.code.sf.net/p/nel/code/trunk@159 bf685bcc-7839-4fcd-9254-36823579bede

+++ nel/tools/3d/zone_welder/zone_welder.cpp
@@ -678,7 +678,7 @@ void weldZones(const char *center)
 				uint	idVert= pa.BaseVertices[corner];
 
 				// write this vertex in array.
-				centerVertices.resize( max(centerVertices.size(), idVert+1) );
+				centerVertices.resize( max((uint)centerVertices.size(), idVert+1) );
 				centerVertices[idVert]= pa.Patch.Vertices[corner];
 			}
 		}
@@ -703,7 +703,7 @@ void weldZones(const char *center)
 					uint	idVert= pa.BaseVertices[corner];
 
 					// write this vertex in array.
-					verts.resize( max(verts.size(), idVert+1) );
+					verts.resize( max((uint)verts.size(), idVert+1) );
 					verts[idVert].Vertex= pa.Patch.Vertices[corner];
 					verts[idVert].OnBorder= false;
 				}
